requirements.txt
dlib==19.24.4
einops==0.4.1
imgaug==0.4.0
lmdb==1.4.1
lpips==0.1.4
matplotlib==3.5.1
mc==0.0
numpy==1.21.2
opencv_contrib_python==4.1.2.30
opencv_python==4.1.2.30
Pillow==8.4.0
Pillow==10.3.0
pyiqa==0.1.11
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
scikit_image==0.19.1
scipy==1.7.2
seaborn==0.13.2
setuptools==58.0.4
tensorflow_gpu==2.7.0
thop==0.1.1.post2209072238
timm==0.6.12
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
wandb==0.17.0
